% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{display_label_map}
\alias{display_label_map}
\title{Map from data names to human-readable display labels}
\format{
An object of class \code{character} of length 27.
}
\usage{
display_label_map
}
\description{
Map from data names to human-readable display labels
}
\keyword{datasets}
