% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{format_posthoc_summary}
\alias{format_posthoc_summary}
\title{Format post-hoc summary}
\usage{
format_posthoc_summary(posthoc)
}
\arguments{
\item{posthoc}{Post-hoc test results}
}
\value{
Formatted summary message
}
\description{
Format post-hoc summary
}
