% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offp-package.R
\docType{package}
\name{offp-package}
\alias{offp}
\alias{offp-package}
\title{offp: Statistical Analysis Functions for Findlay et al. ("OFF project")}
\description{
Statistical analysis functions and data loading utilities for the Findlay et al. manuscript on OFF periods and local sleep in wake.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grahamfindlay/r-offp}
  \item Report bugs at \url{https://github.com/grahamfindlay/r-offp/issues}
}

}
\author{
\strong{Maintainer}: Graham Findlay \email{gfindlay@wisc.edu}

}
\keyword{internal}
