% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_analysis.R
\docType{data}
\name{PREDICTOR_CONDITION}
\alias{PREDICTOR_CONDITION}
\alias{REBOUND_POST}
\alias{REBOUND_BASELINE}
\title{Default predictor / rebound conditions for the correlation analysis}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
PREDICTOR_CONDITION

REBOUND_POST

REBOUND_BASELINE
}
\description{
\code{PREDICTOR_CONDITION} is the whole-period NOD window used as the x-axis;
\code{REBOUND_POST} minus \code{REBOUND_BASELINE} is the \code{NREM.Rebound} response (the
same \code{NREM.Rebound} contrast defined in cnpix_local_sleep's \code{const.py}). All three are
exported by cnpix_local_sleep's \verb{off-analysis export-nod-rebound-correlation}.
}
\keyword{datasets}
