% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{strip_random_effects}
\alias{strip_random_effects}
\title{Strip random effects from formula}
\usage{
strip_random_effects(formula_obj)
}
\arguments{
\item{formula_obj}{Formula object with random effects}
}
\value{
Formula object with only fixed effects
}
\description{
Reconstruct a formula with only fixed effects.
Example A: y ~ x * z + (1|subject) -> y ~ x * z
Example B: y ~ x + (1|subject) -> y ~ x
Example C: y ~ (1|subject) -> y ~ 1
Example D: y ~ (1|subject) + (1|group) -> y ~ 1
For most formulas, this will be equivalent to nobars(formula_obj).
}
